Table param allows increase control of table data.
I really preferred to use a datetime for the time field rather than int since epoch so this allows you to pass in the table name and the preferred datetime field. It should also make it easier to increase other table specific changes that may be needed in the future.