Backwards compatibility by checking if table-data is an array.
The __construct method signature is still the same, but the table-data parameter allows the passing in of a string or array.

Within the  constructor I check to see if the $table-data param is an array, if so we check for further table-setting data. If not we assume that the table-data is a string and assign it to the name of the table. We also set the default time-type to epoch time.

So now we have two different ways of initialising the class:-

Original format :-
$mySQLHandler = new MySQLHandler($pdo, 'foo_bar_log', array('extra_data',), \Monolog\Logger::DEBUG);

Updated format :-

$table_data = array(table_name'=>'tytn_table', 'time_type' => 'DATETIME');
$mySQLHandler = new MySQLHandler($pdo, $table_data , array('extra_data',), \Monolog\Logger::DEBUG);